+++ projeto-map/myapp/api/v1/serializers/ChangePasswordSerializer.py
@@ -1,0 +1,46 @@
+
+from django.contrib.auth.password_validation import validate_password
+from rest_framework import serializers
+from django.core.exceptions import ValidationError
+class ChangePasswordSerializer(serializers.Serializer):
+    """
+    Serializer para mudança de senha
+    """
+    old_password = serializers.CharField(
+        write_only=True,
+        style={'input_type': 'password'},
+        help_text="Sua senha atual"
+    )
+    new_password = serializers.CharField(
+        write_only=True,
+        min_length=8,
+        style={'input_type': 'password'},
+        help_text="Nova senha (mínimo 8 caracteres)"
+    )
+    new_password_confirm = serializers.CharField(
+        write_only=True,
+        style={'input_type': 'password'},
+        help_text="Confirme a nova senha"
+    )
+    
+    def validate(self, attrs):
+        """
+        Valida se as novas senhas coincidem
+        """
+        new_password = attrs.get('new_password')
+        new_password_confirm = attrs.get('new_password_confirm')
+        
+        if new_password != new_password_confirm:
+            raise serializers.ValidationError({
+                'new_password_confirm': 'As novas senhas não coincidem.'
+            })
+        
+        # Validação da nova senha
+        try:
+            validate_password(new_password)
+        except ValidationError as e:
+            raise serializers.ValidationError({
+                'new_password': list(e.messages)
+            })
+        
+        return attrs
