+++ projeto-map/myapp/api/v1/serializers/ProfileSerializer.py
@@ -1,0 +1,24 @@
+
+from django.contrib.auth.models import User
+from rest_framework import serializers
+
+
+
+
+class UserProfileSerializer(serializers.ModelSerializer):
+    """
+    Serializer para perfil do usuário
+    """
+    class Meta:
+        model = User
+        fields = ('id', 'username', 'email', 'first_name', 'last_name', 'date_joined', 'is_active')
+        read_only_fields = ('id', 'username', 'date_joined', 'is_active')
+    
+    def validate_email(self, value):
+        """
+        Verifica se o email já está em uso por outro usuário
+        """
+        user = self.context['request'].user
+        if User.objects.filter(email=value).exclude(pk=user.pk).exists():
+            raise serializers.ValidationError("Este email já está sendo usado por outro usuário.")
+        return value
